Split inheritance for PostgreSQL client
