+++ cliquet/permission/redis.py
@@ -8,7 +8,7 @@ from cliquet.permission import PermissionBase
 from cliquet.storage.redis import wrap_redis_error
 
 
-class Redis(PermissionBase):
+class Permission(PermissionBase):
     """Permission backend implementation using Redis.
 
     Enable in configuration::
@@ -27,7 +27,7 @@ class Redis(PermissionBase):
     """
 
     def __init__(self, *args, **kwargs):
-        super(Redis, self).__init__(*args, **kwargs)
+        super(Permission, self).__init__(*args, **kwargs)
         maxconn = kwargs.pop('max_connections')
         connection_pool = redis.BlockingConnectionPool(max_connections=maxconn)
         self._client = redis.StrictRedis(connection_pool=connection_pool,
@@ -170,8 +170,8 @@ def load_from_config(config):
     uri = urlparse.urlparse(uri)
     pool_size = int(settings['permission_pool_size'])
 
-    return Redis(max_connections=pool_size,
-                 host=uri.hostname or 'localhost',
-                 port=uri.port or 6739,
-                 password=uri.password or None,
-                 db=int(uri.path[1:]) if uri.path else 0)
+    return Permission(max_connections=pool_size,
+                      host=uri.hostname or 'localhost',
+                      port=uri.port or 6739,
+                      password=uri.password or None,
+                      db=int(uri.path[1:]) if uri.path else 0)
